add nodes/ sub-package, slim __init__.py, remove framepack_helpers.py
