--- framepack_helpers.py
@@ -1,1093 +1,0 @@
-"""
-FramePack Helper Functions
-Contains utility functions for video generation, encoding, context building, and benchmarking.
-"""
-
-import torch
-import torch.nn.functional as F
-import numpy as np
-import math
-import time
-import gc
-import json
-import os
-from typing import List, Tuple, Dict
-from datetime import datetime
-from tqdm import tqdm
-import psutil
-
-from pynvml import (
-    nvmlInit, nvmlShutdown, nvmlDeviceGetCount,
-    nvmlDeviceGetHandleByIndex, nvmlDeviceGetMemoryInfo,
-    nvmlDeviceGetUtilizationRates
-)
-
-from comfy import model_management as mm
-from comfy.utils import common_upscale
-
-from .wanvideo.modules.model import rope_params
-from diffusers.schedulers import FlowMatchEulerDiscreteScheduler, DEISMultistepScheduler
-from .wanvideo.utils.fm_solvers import FlowDPMSolverMultistepScheduler
-from .wanvideo.utils.fm_solvers_unipc import FlowUniPCMultistepScheduler
-from .wanvideo.utils.basic_flowmatch import FlowMatchScheduler
-from .wanvideo.utils.scheduling_flow_match_lcm import FlowMatchLCMScheduler
-
-
-# Constants
-VAE_STRIDE = (4, 8, 8)
-PATCH_SIZE = (1, 2, 2)
-
-
-class BenchmarkManager:
-    """Manages benchmarking and performance tracking"""
-    
-    def __init__(self):
-        self.section_benchmarks = {}
-        self.overall_start_time = None
-        self.generation_params = {}
-    
-    def get_memory_stats(self):
-        """Get current memory statistics"""
-        stats = {}
-
-        # CPU memory (process)
-        process = psutil.Process()
-        stats["cpu_memory_mb"] = process.memory_info().rss / 1024 / 1024
-        stats["cpu_memory_percent"] = process.memory_percent()
-
-        # System memory
-        mem = psutil.virtual_memory()
-        stats["system_memory_total_gb"] = mem.total / 1024**3
-        stats["system_memory_used_gb"] = mem.used / 1024**3
-        stats["system_memory_percent"] = mem.percent
-
-        # GPU stats (if CUDA available)
-        if torch.cuda.is_available():
-            stats["gpu_memory_allocated_gb"] = torch.cuda.memory_allocated() / 1024**3
-            stats["gpu_memory_reserved_gb"] = torch.cuda.memory_reserved() / 1024**3
-
-            try:
-                nvmlInit()
-                device_count = nvmlDeviceGetCount()
-                if device_count > 0:
-                    handle = nvmlDeviceGetHandleByIndex(0)
-                    mem_info = nvmlDeviceGetMemoryInfo(handle)
-                    stats["gpu_memory_total_gb"] = mem_info.total / 1024**3
-                    stats["gpu_memory_used_gb"] = mem_info.used / 1024**3
-                    util = nvmlDeviceGetUtilizationRates(handle)
-                    stats["gpu_utilization_percent"] = util.gpu
-            except Exception as e:
-                print(f"GPU stats error: {e}")
-            finally:
-                try:
-                    nvmlShutdown()
-                except:
-                    pass
-        else:
-            stats.update({
-                "gpu_memory_allocated_gb": 0,
-                "gpu_memory_reserved_gb": 0,
-                "gpu_memory_total_gb": 0,
-                "gpu_memory_used_gb": 0,
-                "gpu_utilization_percent": 0,
-            })
-
-        return stats
-
-    def benchmark_section(self, section_id, phase_name):
-        """Start or end benchmarking for a section phase"""
-        if section_id not in self.section_benchmarks:
-            self.section_benchmarks[section_id] = {}
-        
-        phase_key = f"{phase_name}_start"
-        phase_end_key = f"{phase_name}_end"
-        
-        if phase_key not in self.section_benchmarks[section_id]:
-            # Starting phase
-            self.section_benchmarks[section_id][phase_key] = time.time()
-            self.section_benchmarks[section_id][f"{phase_name}_memory_start"] = self.get_memory_stats()
-        else:
-            # Ending phase
-            self.section_benchmarks[section_id][phase_end_key] = time.time()
-            self.section_benchmarks[section_id][f"{phase_name}_memory_end"] = self.get_memory_stats()
-            
-            # Calculate duration
-            duration = self.section_benchmarks[section_id][phase_end_key] - self.section_benchmarks[section_id][phase_key]
-            self.section_benchmarks[section_id][f"{phase_name}_duration"] = duration
-            
-            # Calculate memory delta
-            start_mem = self.section_benchmarks[section_id][f"{phase_name}_memory_start"]
-            end_mem = self.section_benchmarks[section_id][f"{phase_name}_memory_end"]
-            
-            if 'gpu_memory_allocated_gb' in start_mem and 'gpu_memory_allocated_gb' in end_mem:
-                gpu_delta = end_mem['gpu_memory_allocated_gb'] - start_mem['gpu_memory_allocated_gb']
-                self.section_benchmarks[section_id][f"{phase_name}_gpu_memory_delta_gb"] = gpu_delta
-            
-            cpu_delta = end_mem['cpu_memory_mb'] - start_mem['cpu_memory_mb']
-            self.section_benchmarks[section_id][f"{phase_name}_cpu_memory_delta_mb"] = cpu_delta
-
-    def log_metric(self, section_id, metric_name, value):
-        """Log a custom quality metric for a section"""
-        if section_id not in self.section_benchmarks:
-            self.section_benchmarks[section_id] = {}
-        self.section_benchmarks[section_id][metric_name] = value
-    
-    def generate_report(self, section_prompts=None):
-        """Generate a comprehensive benchmark report"""
-        report = []
-        report.append("=" * 80)
-        report.append("FRAMEPACK VIDEO GENERATION BENCHMARK REPORT")
-        report.append("=" * 80)
-        report.append(f"Generated at: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
-        report.append("")
-        
-        if self.overall_start_time:
-            total_duration = time.time() - self.overall_start_time
-            report.append(f"Total Processing Time: {total_duration:.2f} seconds ({total_duration/60:.2f} minutes)")
-        
-        if self.generation_params:
-            report.append("\nGeneration Parameters:")
-            for key, value in self.generation_params.items():
-                report.append(f"  {key}: {value}")
-        
-        report.append("\n" + "-" * 80)
-        report.append("SECTION-BY-SECTION BREAKDOWN")
-        report.append("-" * 80)
-        
-        total_encoding_time = 0
-        total_denoising_time = 0
-        total_accumulation_time = 0
-        
-        for section_id in sorted(self.section_benchmarks.keys()):
-            section_data = self.section_benchmarks[section_id]
-            report.append(f"\n[Section {section_id + 1}]")
-            
-            # Prompt info
-            if section_prompts and section_id < len(section_prompts):
-                report.append(f"Prompt: {section_prompts[section_id][:50]}...")
-            
-            # Timing breakdown
-            phases = ['encoding', 'denoising', 'accumulation']
-            for phase in phases:
-                if f"{phase}_duration" in section_data:
-                    duration = section_data[f"{phase}_duration"]
-                    report.append(f"  {phase.capitalize()}: {duration:.2f}s")
-                    
-                    if phase == 'encoding':
-                        total_encoding_time += duration
-                    elif phase == 'denoising':
-                        total_denoising_time += duration
-                    elif phase == 'accumulation':
-                        total_accumulation_time += duration
-                    
-                    # Memory changes
-                    if f"{phase}_gpu_memory_delta_gb" in section_data:
-                        gpu_delta = section_data[f"{phase}_gpu_memory_delta_gb"]
-                        report.append(f"    GPU Memory Δ: {gpu_delta:+.3f} GB")
-                    
-                    if f"{phase}_cpu_memory_delta_mb" in section_data:
-                        cpu_delta = section_data[f"{phase}_cpu_memory_delta_mb"]
-                        report.append(f"    CPU Memory Δ: {cpu_delta:+.1f} MB")
-            
-            # Per-section total
-            section_total = sum([section_data.get(f"{p}_duration", 0) for p in phases])
-            report.append(f"  Section Total: {section_total:.2f}s")
-            
-            # Peak memory for section
-            if 'denoising_memory_end' in section_data:
-                end_mem = section_data['denoising_memory_end']
-                if 'gpu_memory_allocated_gb' in end_mem:
-                    report.append(f"  Peak GPU Memory: {end_mem['gpu_memory_allocated_gb']:.3f} GB")
-                report.append(f"  Peak CPU Memory: {end_mem['cpu_memory_mb']:.1f} MB")
-        
-        # Summary statistics
-        report.append("\n" + "=" * 80)
-        report.append("SUMMARY STATISTICS")
-        report.append("=" * 80)
-        
-        num_sections = len(self.section_benchmarks)
-        report.append(f"Total Sections Processed: {num_sections}")
-        report.append(f"Total Encoding Time: {total_encoding_time:.2f}s")
-        report.append(f"Total Denoising Time: {total_denoising_time:.2f}s")
-        report.append(f"Total Accumulation Time: {total_accumulation_time:.2f}s")
-        
-        if num_sections > 0:
-            report.append(f"Average Time per Section: {(total_encoding_time + total_denoising_time + total_accumulation_time) / num_sections:.2f}s")
-            report.append(f"Average Denoising Time per Section: {total_denoising_time / num_sections:.2f}s")
-        
-        # Final memory state
-        final_memory = self.get_memory_stats()
-        report.append(f"\nFinal Memory State:")
-        if 'gpu_memory_allocated_gb' in final_memory:
-            report.append(f"  GPU Memory: {final_memory['gpu_memory_allocated_gb']:.3f} GB allocated")
-        report.append(f"  CPU Memory: {final_memory['cpu_memory_mb']:.1f} MB")
-        report.append(f"  System Memory: {final_memory['system_memory_percent']:.1f}% used")
-        
-        report.append("\n" + "=" * 80)
-        
-        return "\n".join(report)
-    
-    def save_report(self, report, output_dir="./benchmarks"):
-        """Save benchmark report to file"""
-        os.makedirs(output_dir, exist_ok=True)
-        
-        timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
-        filename = f"framepack_benchmark_{timestamp}.txt"
-        filepath = os.path.join(output_dir, filename)
-        
-        with open(filepath, 'w') as f:
-            f.write(report)
-        
-        # Also save as JSON for easier analysis
-        json_filename = f"framepack_benchmark_{timestamp}.json"
-        json_filepath = os.path.join(output_dir, json_filename)
-        
-        benchmark_dict = {
-            'timestamp': timestamp,
-            'generation_params': self.generation_params,
-            'section_benchmarks': self.section_benchmarks,
-            'total_duration': time.time() - self.overall_start_time if self.overall_start_time else 0
-        }
-        
-        with open(json_filepath, 'w') as f:
-            json.dump(benchmark_dict, f, indent=2, default=str)
-        
-        print(f"Benchmark report saved to: {filepath}")
-        print(f"JSON data saved to: {json_filepath}")
-        
-        return filepath
-
-
-class PromptHandler:
-    """Handles prompt parsing and encoding"""
-    
-    @staticmethod
-    def parse_multi_prompts(multi_prompts, num_sections):
-        """
-        Parse multi-line prompts and assign them to sections.
-        Each line represents a prompt for a section.
-        If fewer prompts than sections, the last prompt is repeated.
-        """
-        # Split by newline and filter empty lines
-        prompts = [p.strip() for p in multi_prompts.split('\n') if p.strip()]
-        
-        if not prompts:
-            raise ValueError("No prompts provided in multi_prompts")
-        
-        # Assign prompts to sections
-        section_prompts = []
-        for section in range(num_sections):
-            if section < len(prompts):
-                section_prompts.append(prompts[section])
-            else:
-                # Repeat the last prompt for remaining sections
-                section_prompts.append(prompts[-1])
-        
-        print(f"Parsed {len(prompts)} unique prompts for {num_sections} sections")
-        for i, prompt in enumerate(section_prompts):
-            print(f"  Section {i}: {prompt[:50]}...")
-        
-        return section_prompts
-    
-    @staticmethod
-    def parse_prompt_weights(prompt):
-        """
-        Parse prompt weights in the format (text:weight).
-        Returns cleaned prompt and weight dictionary.
-        """
-        import re
-        
-        weights = {}
-        cleaned_prompt = prompt
-        
-        # Pattern to find (text:weight) format
-        pattern = r'\(([^:)]+):([0-9.]+)\)'
-        matches = re.findall(pattern, prompt)
-        
-        for text, weight_str in matches:
-            try:
-                weight = float(weight_str)
-                weights[text.strip()] = weight
-                # Remove the weight notation from the prompt
-                cleaned_prompt = cleaned_prompt.replace(f"({text}:{weight_str})", text)
-            except ValueError:
-                print(f"Invalid weight value: {weight_str}")
-        
-        return cleaned_prompt.strip(), weights
-    
-    @staticmethod
-    def encode_prompt_for_section(prompt, negative_prompt, text_encoder, device):
-        """
-        Encode a single prompt for a section using the WAN Video text encoder.
-        Supports weighted prompts using (text:weight) syntax.
-        """
-        if text_encoder is None:
-            raise ValueError("Text encoder is required for encoding prompts")
-        
-        # Extract the encoder model and dtype
-        encoder = text_encoder["model"]
-        dtype = text_encoder["dtype"]
-        
-        # Split positive prompts by '|' and process weights
-        positive_prompts_raw = [p.strip() for p in prompt.split('|')]
-        positive_prompts = []
-        all_weights = []
-        
-        for p in positive_prompts_raw:
-            cleaned_prompt, weights = PromptHandler.parse_prompt_weights(p)
-            positive_prompts.append(cleaned_prompt)
-            all_weights.append(weights)
-        
-        # Move encoder to device
-        encoder.model.to(device)
-        
-        try:
-            with torch.autocast(device_type=mm.get_autocast_device(device), dtype=dtype, enabled=True):
-                # Encode positive and negative prompts
-                context = encoder(positive_prompts, device)
-                context_null = encoder([negative_prompt if negative_prompt else ""], device)
-                
-                # Apply weights to embeddings if any were extracted
-                for i, weights in enumerate(all_weights):
-                    if weights:  # Only apply if weights exist
-                        for text, weight in weights.items():
-                            print(f"Applying weight {weight} to prompt: {text}")
-                            context[i] = context[i] * weight
-        finally:
-            # Always move encoder back to CPU to free VRAM
-            encoder.model.to('cpu')
-            mm.soft_empty_cache()
-        
-        # Create the embedding dictionary with all required fields for WAN Video
-        prompt_embeds_dict = {
-            "prompt_embeds": context,
-            "negative_prompt_embeds": context_null,
-        }
-        
-        return prompt_embeds_dict
-
-
-class SchedulerFactory:
-    """Factory for creating schedulers"""
-    
-    @staticmethod
-    def create_scheduler(scheduler_name, steps, shift, device, sigmas=None):
-        """Setup the appropriate scheduler"""
-        
-        if scheduler_name == "dpm++":
-            scheduler = FlowDPMSolverMultistepScheduler(shift=shift, algorithm_type="dpmsolver++")
-            if sigmas is None:
-                scheduler.set_timesteps(steps, device=device)
-            else:
-                scheduler.sigmas = sigmas.to(device)
-                scheduler.timesteps = (scheduler.sigmas[:-1] * 1000).to(torch.int64).to(device)
-                scheduler.num_inference_steps = len(scheduler.timesteps)
-                
-        elif scheduler_name == "unipc":
-            scheduler = FlowUniPCMultistepScheduler(shift=shift)
-            if sigmas is None:
-                scheduler.set_timesteps(steps, device=device, shift=shift)
-            else:
-                scheduler.sigmas = sigmas.to(device)
-                scheduler.timesteps = (scheduler.sigmas[:-1] * 1000).to(torch.int64).to(device)
-                scheduler.num_inference_steps = len(scheduler.timesteps)
-                
-        elif scheduler_name == "euler":
-            scheduler = FlowMatchEulerDiscreteScheduler(shift=shift)
-            scheduler.set_timesteps(steps, device=device, sigmas=sigmas.tolist() if sigmas else None)
-            
-        elif scheduler_name == "deis":
-            scheduler = DEISMultistepScheduler(
-                use_flow_sigmas=True,
-                prediction_type="flow_prediction",
-                flow_shift=shift
-            )
-            scheduler.set_timesteps(steps, device=device)
-            scheduler.sigmas[-1] = 1e-6
-            
-        elif scheduler_name == "lcm":
-            scheduler = FlowMatchLCMScheduler(shift=shift)
-            scheduler.set_timesteps(steps, device=device, sigmas=sigmas.tolist() if sigmas else None)
-            
-        else:
-            raise ValueError(f"Unknown scheduler: {scheduler_name}")
-        
-        return scheduler
-
-
-class RoPEEmbeddings:
-    """Handles RoPE embeddings setup"""
-    
-    @staticmethod
-    def setup_rope_embeddings(model_wrapper, latent_video_length):
-        """Setup RoPE embeddings for the model"""
-        
-        model_wrapper.rope_embedder.k = None
-        model_wrapper.rope_embedder.num_frames = None
-        
-        d = model_wrapper.dim // model_wrapper.num_heads
-        riflex_freq_index = 0
-        
-        freqs = torch.cat([
-            rope_params(1024, d - 4 * (d // 6), L_test=latent_video_length, k=riflex_freq_index),
-            rope_params(1024, 2 * (d // 6)),
-            rope_params(1024, 2 * (d // 6))
-        ], dim=1)
-        
-        return freqs
-
-
-class VAEProcessor:
-    """Handles VAE encoding and decoding operations"""
-    
-    def __init__(self, vae, device):
-        self.vae = vae
-        self.device = device
-    
-    def encode_frames(self, frames, ref_images, masks=None, tiled_vae=False):
-        """Encode frames to latent space"""
-        print(f"VAE Encoding debug:")
-        print(f"  Input frames type: {type(frames)}")
-        if isinstance(frames, list):
-             print(f"  Input frames list len: {len(frames)}")
-             if len(frames) > 0:
-                 print(f"  Frame 0 shape: {frames[0].shape}")
-        elif isinstance(frames, torch.Tensor):
-             print(f"  Input frames shape: {frames.shape}")
-             
-        if ref_images is None:
-            ref_images = [None] * len(frames)
-        else:
-            assert len(frames) == len(ref_images)
-
-        if masks is None:
-            latents = self.vae.encode(frames, device=self.device, tiled=tiled_vae)
-        else:
-            inactive = [i * (1 - m) + 0 * m for i, m in zip(frames, masks)]
-            reactive = [i * m + 0 * (1 - m) for i, m in zip(frames, masks)]
-            inactive = self.vae.encode(inactive, device=self.device, tiled=tiled_vae)
-            reactive = self.vae.encode(reactive, device=self.device, tiled=tiled_vae)
-            latents = [torch.cat((u, c), dim=0) for u, c in zip(inactive, reactive)]
-        
-        self.vae.model.clear_cache()
-        cat_latents = []
-        
-        for latent, refs in zip(latents, ref_images):
-            if refs is not None:
-                if masks is None:
-                    ref_latent = self.vae.encode(refs, device=self.device, tiled=tiled_vae)
-                else:
-                    ref_latent = self.vae.encode(refs, device=self.device, tiled=tiled_vae)
-                    ref_latent = [torch.cat((u, torch.zeros_like(u)), dim=0) for u in ref_latent]
-
-                latent = torch.cat([*ref_latent, latent], dim=1)
-            cat_latents.append(latent)
-        
-        return cat_latents
-
-    def encode_masks(self, masks, ref_images=None):
-        """Encode masks to latent space"""
-        if ref_images is None:
-            ref_images = [None] * len(masks)
-        else:
-            assert len(masks) == len(ref_images)
-
-        result_masks = []
-        for mask, refs in zip(masks, ref_images):
-            c, depth, height, width = mask.shape
-            new_depth = int((depth + 3) // VAE_STRIDE[0])
-            height = 2 * (int(height) // (VAE_STRIDE[1] * 2))
-            width = 2 * (int(width) // (VAE_STRIDE[2] * 2))
-
-            # reshape
-            mask = mask[0, :, :, :]
-            mask = mask.view(
-                depth, height, VAE_STRIDE[1], width, VAE_STRIDE[1]
-            )
-            mask = mask.permute(2, 4, 0, 1, 3)
-            mask = mask.reshape(
-                VAE_STRIDE[1] * VAE_STRIDE[2], depth, height, width
-            )
-
-            # interpolation
-            mask = F.interpolate(mask.unsqueeze(0), size=(new_depth, height, width), mode='nearest-exact').squeeze(0)
-
-            if refs is not None:
-                # Calculate latent temporal dimension for reference images
-                # refs shape is [1, C, T, H, W]
-                pixel_frames = refs.shape[2]
-                latent_ref_length = (pixel_frames - 1) // VAE_STRIDE[0] + 1
-                
-                # Create zero mask for reference frames
-                mask_pad = torch.zeros(
-                    (mask.shape[0], latent_ref_length, mask.shape[2], mask.shape[3]),
-                    device=mask.device,
-                    dtype=mask.dtype
-                )
-                mask = torch.cat((mask_pad, mask), dim=1)
-            result_masks.append(mask)
-        
-        return result_masks
-    
-    def combine_latent(self, z, m):
-        """Combine latents and masks"""
-        return [torch.cat([zz, mm], dim=0) for zz, mm in zip(z, m)]
-
-    def decode_latent(self, zs, ref_images=None):
-        """Decode latents back to frames"""
-        return self.vae.decode(zs, device=self.device)
-
-    def decode_single_frame(self, latent, index=-1):
-        """Decode a specific frame index from a latent tensor [C, T, H, W]"""
-        # latent is usually [C, T, H, W]
-        # We need to wrap it for the VAE which expects [B, C, T, H, W] or list
-        # FramePack VAE expects a list of tensors
-        single_latent = latent[:, index:index+1, :, :].clone()
-        return self.vae.decode([single_latent], device=self.device)
-
-
-class ContextBuilder:
-    """Handles hierarchical context building and frame selection"""
-    
-    @staticmethod
-    def build_hierarchical_context(accumulated_latents, section_id):
-        """Build hierarchical context from accumulated latents"""
-        if not accumulated_latents:
-            raise ValueError("No accumulated latents available")
-
-        all_prev = torch.cat(accumulated_latents, dim=1)
-        total_frames = all_prev.shape[1]
-
-        print(f"Building context from {total_frames} accumulated frames")
-
-        return all_prev
-    
-    @staticmethod
-    def pick_context(frames, section_id, initial=False):
-        """
-        Enhanced hierarchical context selection with constant 41-frame output.
-        """
-        # Constants
-        LONG_FRAMES = 14
-        MID_FRAMES = 8
-        RECENT_FRAMES = 3
-        OVERLAP_FRAMES = 5
-        GEN_FRAMES = 30
-        TOTAL_FRAMES = 60
-
-        C, T, H, W = frames.shape
-
-        if initial and T == TOTAL_FRAMES:
-            return frames
-
-        if initial and T < TOTAL_FRAMES:
-            padding_needed = TOTAL_FRAMES - T
-            padding = torch.zeros((C, padding_needed, H, W), device=frames.device)
-            return torch.cat([frames, padding], dim=1)
-
-        # SIMPLIFIED STRATEGY: Use the last 11 frames contiguously
-        CONTEXT_FRAMES = LONG_FRAMES + MID_FRAMES + RECENT_FRAMES + OVERLAP_FRAMES # 11
-        
-        # Ensure we have enough frames
-        if T < CONTEXT_FRAMES:
-             # Fallback if not enough frames (unlikely after section 0)
-             context_frames = frames
-             padding = torch.zeros((C, CONTEXT_FRAMES - T, H, W), device=frames.device)
-             context_frames = torch.cat([padding, context_frames], dim=1)
-        else:
-             context_frames = frames[:, -CONTEXT_FRAMES:, :, :]
-
-        # Return ONLY the context frames, do not pad with placeholder yet
-        # We will handle padding in pixel space after decoding
-        final_frames = context_frames
-
-        if section_id % 5 == 0 or True: # Always print for now
-            print(f"\nContext selection debug (section {section_id}):")
-            print(f"  Input frames: {T}")
-            print(f"  Strategy: Contiguous last {CONTEXT_FRAMES} frames")
-            print(f"  Output shape: {final_frames.shape}")
-
-        return final_frames
-
-
-class MaskGenerator:
-    """Handles mask generation for temporal blending"""
-    
-    @staticmethod
-    def create_temporal_blend_mask(frame_shape, section_id, device, initial=False):
-        """Enhanced mask creation that handles decoded frame dimensions"""
-        C, T, H, W = frame_shape
-        
-        # Constants
-        LATENT_FRAMES = 60
-        decoded_frames = T
-        expansion_ratio = decoded_frames / LATENT_FRAMES
-        
-        mask = torch.zeros(3, decoded_frames, H, W, device=device)
-        
-        # Scale all frame counts by the expansion ratio
-        LONG_FRAMES = int(14 * expansion_ratio)
-        MID_FRAMES = int(8 * expansion_ratio)
-        RECENT_FRAMES = int(3 * expansion_ratio)
-        OVERLAP_FRAMES = int(5 * expansion_ratio)
-        GEN_FRAMES = decoded_frames - (LONG_FRAMES + MID_FRAMES + RECENT_FRAMES + OVERLAP_FRAMES)
-        
-        print(f"\nMask generation debug (section {section_id}):")
-        print(f"  Decoded frames: {decoded_frames}")
-        print(f"  Expansion ratio: {expansion_ratio}")
-        print(f"  Segments: L={LONG_FRAMES}, M={MID_FRAMES}, R={RECENT_FRAMES}, O={OVERLAP_FRAMES}, Gen={GEN_FRAMES}")
-        
-        if initial:
-            mask[:, :-GEN_FRAMES] = 0.0
-            mask[:, -GEN_FRAMES:] = 1.0
-            return [mask]
-        
-        # Apply mask values
-        idx = 0
-        mask[:, idx:idx+LONG_FRAMES] = 0.05
-        idx += LONG_FRAMES
-        
-        mask[:, idx:idx+MID_FRAMES] = 0.2
-        idx += MID_FRAMES
-        
-        mask[:, idx:idx+RECENT_FRAMES] = 0.3
-        idx += RECENT_FRAMES
-        
-        for i in range(OVERLAP_FRAMES):
-            blend_value = 0.4 + (i / (OVERLAP_FRAMES - 1)) * 0.4
-            mask[:, idx+i] = blend_value
-        idx += OVERLAP_FRAMES
-        
-        mask[:, idx:] = 1.0
-        
-        return [mask]
-    
-    @staticmethod
-    def create_spatial_variation(H, W, device):
-        """Create spatial variation mask for natural blending"""
-        y_coords = torch.linspace(-1, 1, H, device=device)
-        x_coords = torch.linspace(-1, 1, W, device=device)
-        y_grid, x_grid = torch.meshgrid(y_coords, x_coords, indexing='ij')
-
-        distance = torch.sqrt(x_grid**2 + y_grid**2) / 1.414
-        variation = 1.0 - 0.3 * torch.exp(-3 * distance**2)
-
-        return variation
-
-
-class FrequencyProcessor:
-    """Handles frequency domain operations"""
-    
-    @staticmethod
-    def separate_appearance_and_motion(frames):
-        """Use frequency domain to separate appearance from motion"""
-        C, T, H, W = frames.shape
-
-        # FFT
-        fft_frames = torch.fft.rfft2(frames, dim=(-2, -1))
-        fft_h = H
-        fft_w = W // 2 + 1
-
-        h_freqs = torch.fft.fftfreq(H, device=frames.device)
-        w_freqs = torch.fft.rfftfreq(W, device=frames.device)
-        h_grid, w_grid = torch.meshgrid(h_freqs, w_freqs, indexing='ij')
-
-        freq_magnitude = torch.sqrt(h_grid**2 + w_grid**2)
-        cutoff = 0.1
-        low_pass_mask = (freq_magnitude < cutoff).float().to(frames.device)
-
-        if low_pass_mask.shape != fft_frames.shape[-2:]:
-            low_pass_mask = low_pass_mask[:fft_h, :fft_w]
-
-        while low_pass_mask.dim() < fft_frames.dim():
-            low_pass_mask = low_pass_mask.unsqueeze(0)
-
-        appearance_fft = fft_frames * low_pass_mask
-        motion_fft = fft_frames * (1 - low_pass_mask)
-
-        appearance = torch.fft.irfft2(appearance_fft, s=(H, W))
-        motion = torch.fft.irfft2(motion_fft, s=(H, W))
-
-        return appearance, motion
-
-
-class ReferenceImageProcessor:
-    """Handles reference image processing"""
-    
-    @staticmethod
-    def process_reference_images(ref_images, width, height, device, dtype, target_ref_count=1):
-        """Process reference images for generation"""
-        
-        # Handle duplication if requested
-        if ref_images.shape[0] == 1 and target_ref_count > 1:
-            print(f"Duplicating single reference image {target_ref_count} times")
-            ref_images = ref_images.repeat(target_ref_count, 1, 1, 1)
-
-        B, H, W, C = ref_images.shape
-        current_aspect = W / H
-        target_aspect = width / height
-        
-        if current_aspect > target_aspect:
-            new_h = int(W / target_aspect)
-            pad_h = (new_h - H) // 2
-            padded = torch.ones(ref_images.shape[0], new_h, W, ref_images.shape[3], 
-                            device=ref_images.device, dtype=ref_images.dtype)
-            padded[:, pad_h:pad_h+H, :, :] = ref_images
-            ref_images = padded
-        elif current_aspect < target_aspect:
-            new_w = int(H * target_aspect)
-            pad_w = (new_w - W) // 2
-            padded = torch.ones(ref_images.shape[0], H, new_w, ref_images.shape[3], 
-                            device=ref_images.device, dtype=ref_images.dtype)
-            padded[:, :, pad_w:pad_w+W, :] = ref_images
-            ref_images = padded
-            
-        ref_images = common_upscale(ref_images.movedim(-1, 1), width, height, 
-                                "lanczos", "center")
-        ref_images = ref_images.movedim(0, 1) # [C, T, H, W]
-        ref_images = (ref_images.to(dtype).to(device) * 2 - 1)
-        
-        return [ref_images.unsqueeze(0)]
-
-
-class FramePackCompressor:
-    """
-    Handles hierarchical compression of latent history in latent space.
-    Recent history stays high-res; deep history is pooled.
-    """
-    def __init__(self, 
-                 lambda_compression: float = 2.0,
-                 max_history_frames: int = 120):
-        
-        self.lambda_compression = lambda_compression
-        self.max_history_frames = max_history_frames
-        
-        # Kernel sizes for different levels of compression
-        self.base_kernels = [
-            (1, 1, 1),   # Level 0: 1x
-            (1, 2, 2),   # Level 1: 4x
-            (2, 2, 2),   # Level 2: 8x
-            (2, 4, 4),   # Level 3: 32x
-            (4, 8, 8),   # Level 4: 256x
-        ]
-
-    def _get_kernel_for_section(self, section_age: int) -> Tuple[int, int, int]:
-        """Calculates kernel size based on section age and lambda"""
-        if section_age == 0:
-            return (1, 1, 1)
-        
-        # Exponential growth of compression: lambda^(age)
-        target_rate = self.lambda_compression ** section_age
-        
-        # Find best kernel from our presets
-        best_kernel = self.base_kernels[0]
-        for kernel in self.base_kernels:
-            rate = kernel[0] * kernel[1] * kernel[2]
-            if rate <= target_rate:
-                best_kernel = kernel
-            else:
-                break
-        return best_kernel
-
-    def compress_latent(self, latent: torch.Tensor, kernel: Tuple[int, int, int]) -> torch.Tensor:
-        """Applies avg_pool3d to compress latent dimensions"""
-        if kernel == (1, 1, 1):
-            return latent
-        
-        # latent is [C, T, H, W]
-        # avg_pool3d expects [N, C, T, H, W]
-        with torch.no_grad():
-            compressed = F.avg_pool3d(
-                latent.unsqueeze(0),
-                kernel_size=kernel,
-                stride=kernel
-            ).squeeze(0)
-        return compressed
-
-    def prepare_context(self, accumulated_latents: List[torch.Tensor], section_id: int) -> torch.Tensor:
-        """
-        Builds a single context tensor from the list of historical latent sections.
-        Sections are compressed according to their age.
-        """
-        if not accumulated_latents:
-            return None
-            
-        # Reverse history so newest is index 0
-        history = list(reversed(accumulated_latents))
-        
-        processed_sections = []
-        for age, section in enumerate(history):
-            kernel = self._get_kernel_for_section(age)
-            compressed = self.compress_latent(section, kernel)
-            processed_sections.append(compressed)
-            
-            # Limit history to prevent runaway sequence length
-            if len(processed_sections) >= self.max_history_frames // 10:
-                break
-        
-        # Concatenate into one temporal block
-        # Note: Since they have different H, W after pooling, we pad to match the largest
-        max_h = max(s.shape[2] for s in processed_sections)
-        max_w = max(s.shape[3] for s in processed_sections)
-        
-        padded = []
-        for s in processed_sections:
-            if s.shape[2] < max_h or s.shape[3] < max_w:
-                pad_h = max_h - s.shape[2]
-                pad_w = max_w - s.shape[3]
-                s = F.pad(s, (0, pad_w, 0, pad_h))
-            padded.append(s)
-            
-        return torch.cat(padded, dim=1)
-
-
-class SparseSelector:
-    """
-    Selects frames with exponential backoff to preserve long-term identity.
-    Example: [Last 5 frames, then T-10, T-30, T-60, Frame 0]
-    """
-    @staticmethod
-    def pick_sparse_context(accumulated_latents: List[torch.Tensor], num_frames: int = 30) -> torch.Tensor:
-        if not accumulated_latents:
-            return None
-            
-        all_frames = torch.cat(accumulated_latents, dim=1)
-        total_t = all_frames.shape[1]
-        
-        if total_t <= num_frames:
-            return all_frames
-            
-        # Select indices
-        indices = []
-        
-        # 1. Always include the very first frame (Identity Anchor)
-        indices.append(0)
-        
-        # 2. Always include the last N/2 frames (Continuity)
-        recent_count = num_frames // 2
-        indices.extend(range(total_t - recent_count, total_t))
-        
-        # 3. Fill the rest with exponential spacing from the past
-        remaining = num_frames - len(indices)
-        if remaining > 0:
-            # Calculate points between Frame 1 and (Total - Recent)
-            search_end = total_t - recent_count - 1
-            if search_end > 1:
-                # Use log space for selection
-                log_points = np.linspace(np.log(1), np.log(search_end), remaining)
-                sparse_indices = np.exp(log_points).astype(int)
-                indices.extend(sparse_indices.tolist())
-        
-        # Unique and sorted
-        final_indices = sorted(list(set(indices)))
-        
-        # If we have too many (due to overlaps), take the most important ones
-        if len(final_indices) > num_frames:
-            final_indices = final_indices[-num_frames:]
-            
-        return all_frames[:, final_indices, :, :]
-
-
-class MoCRouter:
-    """
-    Mixture of Contexts Router (Outer Loop).
-    Retrieves historical chunks based on semantic similarity to the current prompt.
-    """
-    @staticmethod
-    def retrieve_context(accumulated_latents: List[torch.Tensor], 
-                         current_prompt_embeds: torch.Tensor,
-                         top_k: int = 3) -> torch.Tensor:
-        """
-        current_prompt_embeds: [1, L, D]
-        returns: Concatenated top-k most similar latent chunks.
-        """
-        if not accumulated_latents:
-            return None
-            
-        if len(accumulated_latents) <= top_k:
-            return torch.cat(accumulated_latents, dim=1)
-            
-        # 1. Prepare Query: Mean pool the prompt embeddings
-        # current_prompt_embeds is usually a list or dict in this node
-        if isinstance(current_prompt_embeds, dict):
-            query = current_prompt_embeds["prompt_embeds"]
-        else:
-            query = current_prompt_embeds
-            
-        # If it's a list, take the first one
-        if isinstance(query, list):
-            query = query[0]
-            
-        # query shape: [1, seq_len, dim] -> [dim]
-        query_vec = query.mean(dim=1).flatten()
-        
-        # 2. Prepare Keys: Mean pool each historical chunk
-        chunk_scores = []
-        for i, chunk in enumerate(accumulated_latents):
-            # chunk shape: [C, T, H, W] -> [C]
-            key_vec = chunk.mean(dim=(1, 2, 3)).flatten()
-            
-            # 3. Calculate Cosine Similarity
-            # Note: Wan Video latents (16 channels) and T5 embeds (4096 dim) 
-            # don't match directly. We use high-level spatial correlations.
-            
-            # ZERO-SHOT TRANSLATION:
-            # Inner Loop: Current Query (Q) vs Historical Keys (K)
-            # Outer Loop (Ours): Last Frame Feature (Query Ref) vs Historical Chunk Means (Keys)
-            
-            # Query Ref: Mean-pool the very last frame of the most recent chunk
-            # accumulated_latents[-1] is [C, T, H, W] -> [C, -1, :, :] -> [C]
-            ref_frame_vec = accumulated_latents[-1][:, -1, :, :].mean(dim=(1, 2)).flatten()
-            
-            similarity = F.cosine_similarity(ref_frame_vec.unsqueeze(0), key_vec.unsqueeze(0))
-            chunk_scores.append((i, similarity.item()))
-            
-        # 4. Select Top-K
-        chunk_scores.sort(key=lambda x: x[1], reverse=True)
-        selected_indices = sorted([x[0] for x in chunk_scores[:top_k]])
-        
-        # 5. Always include the most recent chunk for continuity
-        recent_idx = len(accumulated_latents) - 1
-        if recent_idx not in selected_indices:
-            selected_indices[-1] = recent_idx # Replace least similar with most recent
-            selected_indices.sort()
-            
-        selected_chunks = [accumulated_latents[i] for i in selected_indices]
-        return torch.cat(selected_chunks, dim=1)
-
-
-class VideoMetrics:
-    """Calculates quality and consistency metrics for video generation."""
-
-    @staticmethod
-    def calculate_ssim_boundary(frame_prev: torch.Tensor, frame_next: torch.Tensor) -> float:
-        """
-        Calculates a simplified Structural Similarity (SSIM) between two frames.
-        frame_prev, frame_next: [1, 3, 1, H, W] or [1, 3, H, W] or [3, H, W]
-        """
-        # Ensure 3D tensors [3, H, W] for calculation
-        while frame_prev.dim() > 3:
-            if frame_prev.shape[0] == 1: frame_prev = frame_prev.squeeze(0)
-            elif frame_prev.shape[2] == 1: frame_prev = frame_prev.squeeze(2) # [1, 3, 1, H, W] -> [1, 3, H, W]
-            else: break
-            
-        while frame_next.dim() > 3:
-            if frame_next.shape[0] == 1: frame_next = frame_next.squeeze(0)
-            elif frame_next.shape[2] == 1: frame_next = frame_next.squeeze(2)
-            else: break
-            
-        # Fallback squeeze to exactly 3 dims if still more
-        if frame_prev.dim() > 3: frame_prev = frame_prev.view(3, frame_prev.shape[-2], frame_prev.shape[-1])
-        if frame_next.dim() > 3: frame_next = frame_next.view(3, frame_next.shape[-2], frame_next.shape[-1])
-            
-        # Ensure same spatial size
-        if frame_prev.shape[-2:] != frame_next.shape[-2:]:
-            frame_next = torch.nn.functional.interpolate(frame_next, 
-                                                        size=frame_prev.shape[-2:], 
-                                                        mode='bilinear')
-
-        # Basic SSIM terms: Luminance, Contrast, Structure
-        # We use a global mean/var for simplicity (Zero-Shot)
-        mu1 = frame_prev.mean()
-        mu2 = frame_next.mean()
-        sigma1sq = frame_prev.var()
-        sigma2sq = frame_next.var()
-        sigma12 = ((frame_prev - mu1) * (frame_next - mu2)).mean()
-        
-        c1 = (0.01 * 1.0)**2
-        c2 = (0.03 * 1.0)**2
-        
-        ssim = ((2 * mu1 * mu2 + c1) * (2 * sigma12 + c2)) / \
-               ((mu1**2 + mu2**2 + c1) * (sigma1sq + sigma2sq + c2))
-        
-        return float(ssim.item())
-
-    @staticmethod
-    def calculate_embedding_drift(embed1: torch.Tensor, embed2: torch.Tensor) -> float:
-        """Calculates cosine distance between two embeddings."""
-        # embed1, embed2: [D] or [1, D]
-        sim = torch.nn.functional.cosine_similarity(embed1.view(1, -1), 
-                                                   embed2.view(1, -1))
-        return float(1.0 - sim.item())
-
-    @staticmethod
-    def calculate_semantic_alignment(latent: torch.Tensor, prompt_embed: torch.Tensor) -> float:
-        """
-        A rough proxy for how well the latent matches the prompt theme.
-        We compare the pooled latent features with the pooled prompt features.
-        """
-        # Mean pool latent [C, T, H, W] -> [C]
-        z_feat = latent.mean(dim=(1, 2, 3)).view(1, -1)
-        
-        # Mean pool prompt [1, L, D] -> [D]
-        p_feat = prompt_embed.mean(dim=1).view(1, -1)
-        
-        # Since C (16) and D (4096) don't match, we use PCA or Projection in training
-        # ZERO-SHOT FALLBACK: We measure the variance of the latent as a proxy for "activity/energy"
-        # compared to prompt "length/complexity". 
-        # For now, let's return a dummy placeholder until we have a proper projector.
-        return 0.5
-
-
-class BenchmarkAnalyzer:
-    """Consolitates results from multiple runs into a comparison report."""
-    
-    def __init__(self, output_dir="./benchmarks"):
-        self.output_dir = output_dir
-        os.makedirs(output_dir, exist_ok=True)
-        
-    def save_run_data(self, method_name, benchmark_manager):
-        """Save a single run's data to a JSON file"""
-        data = {
-            "method": method_name,
-            "timestamp": datetime.now().strftime('%Y-%m-%d %H:%M:%S'),
-            "params": benchmark_manager.generation_params,
-            "sections": benchmark_manager.section_benchmarks
-        }
-        
-        filename = os.path.join(self.output_dir, f"run_{method_name}.json")
-        with open(filename, 'w') as f:
-            json.dump(data, f, indent=4)
-        print(f"Saved benchmark data for method '{method_name}' to {filename}")
-
-    def generate_comparison_report(self):
-        """Build a Markdown comparison report from all JSON files in the directory"""
-        files = [f for f in os.listdir(self.output_dir) if f.startswith("run_") and f.endswith(".json")]
-        if not files:
-            return "No benchmark data found to compare."
-            
-        all_data = []
-        for file in files:
-            with open(os.path.join(self.output_dir, file), 'r') as f:
-                all_data.append(json.load(f))
-                
-        report = ["# FramePack Context Method Comparison Report\n"]
-        report.append("| Method | Avg. SSIM (Motion) | Identity Drift | VRAM (Max GB) | Speed (FPS) |")
-        report.append("| :--- | :---: | :---: | :---: | :---: |")
-        
-        for run in all_data:
-            method = run["method"]
-            sections = run["sections"]
-            
-            # Aggregate stats
-            ssims = [v["boundary_ssim"] for v in sections.values() if "boundary_ssim" in v]
-            drifts = [v["identity_drift"] for v in sections.values() if "identity_drift" in v]
-            vrams = [v["denoising_memory_end"]["gpu_memory_allocated_gb"] for v in sections.values() if "denoising_memory_end" in v]
-            durations = [v["denoising_duration"] for v in sections.values() if "denoising_duration" in v]
-            
-            avg_ssim = sum(ssims)/len(ssims) if ssims else 0
-            avg_drift = sum(drifts)/len(drifts) if drifts else 0
-            max_vram = max(vrams) if vrams else 0
-            
-            total_frames = run["params"].get("num_frames", 30)
-            total_duration = sum(durations) if durations else 1
-            fps = total_frames / total_duration
-            
-            report.append(f"| {method} | {avg_ssim:.4f} | {avg_drift:.4f} | {max_vram:.2f} | {fps:.2f} |")
-            
-        report_text = "\n".join(report)
-        
-        with open(os.path.join(self.output_dir, "comparison_report.md"), 'w') as f:
-            f.write(report_text)
-            
-        return report_text
